backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
anthropic
python-multipart
python-jose[cryptography]
httpx
python-dotenv
